# frontend/test/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis",
    "description": "Static-scan hits + LSP call graph -> AI-ready analysis bundles. First deliverable: the assembly stage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/scan": {
      "post": {
        "summary": "Scan",
        "operationId": "scan_v1_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOnlyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scan/jobs": {
      "post": {
        "summary": "Scan Job",
        "description": "The asynchronous form of `/v1/scan`. The synchronous route is unchanged.\n\nA separate path rather than a change to `/v1/scan`: that response is a documented\ncontract (`run_id`/`findings`/`sarif_path`/`engine`) with existing consumers, and\nquietly turning it into \"202 and an id\" would be a breaking change disguised as a\nfeature.",
        "operationId": "scan_job_v1_scan_jobs_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanOnlyResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "findings": {
            "type": "integer",
            "title": "Findings"
          },
          "sarif_path": {
            "type": "string",
            "title": "Sarif Path"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "findings",
          "sarif_path",
          "engine"
        ],
        "title": "ScanOnlyResponse"
      },
...
      "ScanRequest": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace"
          },
          "rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules"
          },
          "rule_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Config"
          },
          "include_globs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Include Globs"
          },
          "exclude_globs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Globs"
          }
        },
        "type": "object",
        "required": [
          "workspace"
        ],
        "title": "ScanRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}